LayoutTests/imported/w3c/web-platform-tests/speculation-rules/prefetch/referrer-policy.https_4-last-expected.txt


FAIL with "strict-origin" link referrer policy overriding "no-referrer" of referring page assert_equals: expected (string) "https://localhost:9443/" but got (undefined) undefined
